# scripts/reports/req-ai-codegen-004/openapi.json
{
  "components": {
    "parameters": {
      "XRequestId": {
        "in": "header",
        "name": "X-Request-Id",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "examples": {
              "authFailed": {
                "value": {
                  "code": "E1100",
                  "detail": {
                    "error": "missing Authorization"
                  },
                  "hint": "请提供Authorization: Bearer \u003ctoken\u003e",
                  "message": "AuthFailed",
                  "requestId": "00000000-0000-0000-0000-000000000000",
                  "severity": "warning"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Error"
      },
      "OK": {
        "content": {
          "application/json": {
            "examples": {
              "success": {
                "value": {
                  "code": "OK",
                  "data": {
                    "example": "value"
                  },
                  "message": "success",
                  "requestId": "00000000-0000-0000-0000-000000000000",
                  "severity": "info"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Response"
            }
          }
        },
        "description": "OK"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "enum": [
              "E1000",
              "E1100",
              "E1200",
              "E1300",
              "E1400",
              "E1500",
              "E2000",
              "E2100",
              "E3000",
              "E3100"
            ],
            "type": "string"
          },
          "detail": {
            "type": "object"
          },
          "hint": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "LLMChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "properties": {
                "content": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                }
              },
              "required": [
                "role",
                "content"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "modelId": {
            "type": "string"
          },
          "providerHint": {
            "type": "string"
          },
          "stream": {
            "type": "boolean"
          },
          "temperature": {
            "type": "number"
          },
          "topP": {
            "type": "number"
          }
        },
        "required": [
          "modelId",
          "messages"
        ],
        "type": "object"
      },
      "LLMEmbeddingsRequest": {
        "properties": {
          "input": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "modelId": {
            "type": "string"
          }
        },
        "required": [
          "modelId",
          "input"
        ],
        "type": "object"
      },
      "LLMModerateRequest": {
        "properties": {
          "input": {
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "type": "object"
      },
...
      "Response": {
        "properties": {
          "code": {
            "enum": [
              "OK"
            ],
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "headerPermissions": {
        "in": "header",
        "name": "X-User-Permissions",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Xingzuo API",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/articles": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "List articles"
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Create article"
      }
    },
    "/api/v1/articles/:id": {
      "delete": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Delete article"
      },
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Get article"
      },
      "put": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Update article"
      }
    },
...
    "/api/v1/llm/chat": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Unified LLM chat"
      }
    },
    "/api/v1/llm/embeddings": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMEmbeddingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Generate embeddings"
      }
    },
    "/api/v1/llm/models": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "List LLM models"
      }
    },
    "/api/v1/llm/moderate": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMModerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "Content moderation"
      }
    },
    "/api/v1/llm/providers": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "headerPermissions": []
          }
        ],
        "summary": "List LLM providers"
      }
    },
...
  }
}